Simplify some column names
I think it makes more sense to call the column for the name of the group
simply `name` rather than `group_name`, the column is already in the
`groups` table, its implied that the column `name` will be the name of
the group.

Same applies for changing `permission_string` to `name`, I think it's
more descriptive.