Implement `User#permission?` to check whether a user has a permission
`User#permission?` is called with a string as a parameter and return
true/false depending on whther the `User` has that permissions.

This commit also introduces methods to facilitate `User#permission?`
these are `User#permission_strings` and `Group#permission_strings` these
methods return the array of permission strings that a `User` and `Group`
have.