Add basic edit/update functionality to `Profiles`
Adds both `html` and `js` responses allowing the editting/updating of
`first_name`, `last_name`, & `location` on a users profile.

Introduces the `profile.edit` & `profile.edit.others` permissions.

Removes the `Rail/HttpPositionalArguments` Cop from `rubocop`, this is
due to the fact that we're using `rails-4.2.7.1` and that cop is
intended to catch behaviour that was deprecated in this [PR](https://github.com/rails/rails/pull/18323)
however that deprecation does not affect our current `rails` version.