Update `Procfile.dev`, to no longer rely on env vars
Previously, because I'm a genius (cause I copy pasta'd from the production
Procfile), we were relying on env vars to set the environment and port on which
the server should run. This was a dumb idea and is now fixed.

The server will now explicitly run on port 5000 and in the development
environment.