Add check to prevent remove user from his last group
The `before_destroy` action will check if we are attempting to remove a
user from his final group. And fails the deletion if we are.

Also adds a dependent action on the `User` model that will remove join
records between a `User` & a `Group`.

The reason for using `delete_all` rather than `destroy` is to bypass the
`before_destroy` callback, which would prevent the `GroupUser` record
being destroyed.